(ms) Make sure the child isn't killed by a closed session before      it is able to detach from the tty. (ms) Better __DIE__ handler as suggested by Karl Rune Nilsen in      http://rt.cpan.org/Ticket/Display.html?id=39917.

+++ Daemon.pm
@@ -136,6 +136,10 @@ sub daemonize {
     if( $background ) {
         umask(0);
 
+          # Make sure the child isn't killed when the uses closes the
+          # terminal session before the child detaches from the tty.
+        $SIG{'HUP'} = 'IGNORE';
+
         my $child = fork();
 
         if($child < 0) {
@@ -160,7 +164,8 @@ sub daemonize {
     }
 
     $SIG{__DIE__} = sub { 
-        if(! $^S) {
+          # Make sure it's not an eval{} triggering the handler.
+        if(defined $^S && $^S==0) {
             unlink $pidfile or warn "Cannot remove $pidfile";
         }
     };
